refactor: 달력 컴포넌트 제거 및 달력 페이지 컴포넌트로 이동
